BCN-3: Use immutable sequences
Porting some changes from https://github.com/twitter/scrooge/pull/350

Instead of using a flag, use immutable sequences unconditionally. This
build is only meant to be used temporarily until we upgrade finagle.

Since we cannot change scrooge-core (because other libraries in use
depend on it), the generated code is adopted to add the `.toSeq` call
to convert the output type to a `immutable.Seq`.